refactor: 모든 컨트롤러 매핑을 @Controller 어노테이션으로 변경
