+++ app/src/main/java/com/techcourse/controller/ForwardController.java
@@ -1,0 +1,16 @@
+package com.techcourse.controller;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
+
+@Controller
+public class ForwardController {
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public String getRootPage(final HttpServletRequest request, final HttpServletResponse response) {
+        return "/index.jsp";
+    }
+}
