+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -4,12 +4,15 @@ import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
+import nextstep.web.annotation.Controller;
+import nextstep.web.annotation.RequestMapping;
+import nextstep.web.support.RequestMethod;
 
-public class RegisterController implements Controller {
+@Controller
+public class RegisterController {
 
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    public String registerMember(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
         final var user = new User(2,
                 req.getParameter("account"),
                 req.getParameter("password"),
@@ -17,5 +20,10 @@ public class RegisterController implements Controller {
         InMemoryUserRepository.save(user);
 
         return "redirect:/index.jsp";
+    }
+
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
+    public String getRegisterPage(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
+        return "/register.jsp";
     }
 }
